# src/api/auth/sign-up.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * aegis-backend
 * OpenAPI spec version: 0.0.0
 */
export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  Admin: 'Admin',
  Teacher: 'Teacher',
  Student: 'Student',
} as const;

export interface SignUpRequest {
  identity: string;
  /** @nullable */
  name?: string | null;
  password: string;
  role: UserRole;
}