# src/api/index.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * aegis-backend
 * OpenAPI spec version: 0.0.0
 */
export interface SignInRequest {
  identity: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  Admin: 'Admin',
  Teacher: 'Teacher',
  Student: 'Student',
} as const;

export interface SignUpRequest {
  identity: string;
  /** @nullable */
  name?: string | null;
  password: string;
  role: UserRole;
}